refactor(#39): 이전 다음 버튼 공통 컴포넌트로 변경
