+++ src/pages/recommendation/Subjective.tsx
@@ -1,7 +1,7 @@
 import RecommendationProfiles from '../../components/recommendation/RecommendationProfiles';
 import RecommendationKeywords from '../../components/recommendation/RecommendationKeywords';
 import RecommendationTendency from '../../components/recommendation/RecommendationTendency';
-import BeforeNextButton from '../../components/recommendation/BeforeNextButton';
+import BeforeNextButton from '../../components/common/BeforeNextButton';
 
 const Subjective = () => {
   return (
